move existing functions from snakemake scripts to doubletime tools
